Move loader helper functions to new file
